feat: move Store to App
